clear navbar - nav
