+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/navbar/NavbarComplexLinks.java
@@ -5,7 +5,6 @@ import com.epam.jdi.light.elements.composite.Section;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.JDropdown;
 import com.epam.jdi.light.elements.pageobjects.annotations.locators.UI;
 import com.epam.jdi.light.ui.bootstrap.elements.common.Link;
-import com.epam.jdi.light.ui.bootstrap.elements.complex.Collapse;
 import com.epam.jdi.light.ui.bootstrap.elements.complex.ListGroup;
 
 public class NavbarComplexLinks extends Section {
@@ -18,5 +17,10 @@ public class NavbarComplexLinks extends Section {
             expand = "#navbarDropdownMenuLink",
             list = "a")
     public Dropdown dropdown;
-//    public Collapse dropdown;
+
+    public void selectMenu(String item) {
+        dropdown.show();
+        dropdown.toggle();
+        dropdown.list().select(item);
+    }
 }
