+++ jdi-light-bootstrap-tests/src/main/java/io/github/com/sections/navbar/NavbarSimpleLinks.java
@@ -8,5 +8,4 @@ import com.epam.jdi.light.ui.bootstrap.elements.complex.ListGroup;
 public class NavbarSimpleLinks extends Section {
     @UI(".navbar-brand") public Link brand;
     @UI(".navbar-nav a") public ListGroup listPages;
-//    @UI(".navbar-nav .nav-item") public ListGroup listPages;
 }
